+++ tools/tune_vot.py
@@ -142,7 +142,8 @@ def tune(param):
         if args.visualization and f >= start_frame:  # visualization (skip lost frame)
             if f == 0: cv2.destroyAllWindows()
             if len(gt[f]) == 8:
-                cv2.polylines(im, [np.array(gt[f], np.int).reshape((-1, 1, 2))], True, (0, 255, 0), 3)
+                #cv2.polylines(im, [np.array(gt[f], np.int).reshape((-1, 1, 2))], True, (0, 255, 0), 3)
+                cv2.polylines(im, [np.array(gt[f], int).reshape((-1, 1, 2))], True, (0, 255, 0), 3)
             else:
                 cv2.rectangle(im, (gt[f, 0], gt[f, 1]), (gt[f, 0] + gt[f, 2], gt[f, 1] + gt[f, 3]), (0, 255, 0), 3)
             if len(location) == 8:
